% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryspellapprastviz.R
\name{dryspellapprastviz}
\alias{dryspellapprastviz}
\title{Heat and Cold waves analysis in Spatial Gridded Coverage (visualization)}
\usage{
dryspellapprastviz(
  x,
  filenames,
  sf,
  settings = system.file("settings/lm_plot_settings_v4.xml", package = "terraclivaviz"),
  fun_aggr = terracliva::aggr_fun_suffixes(),
  summary_suffixes = c("pvalue", "coeff", "stdrerror", "rsquared", "senslope",
    "pvalue_mk"),
  ...
)
}
\arguments{
\item{x}{a \code{SpatRast-Class} object returned by \code{\link{dryspellapprast}}.}

\item{filenames}{vector or string for names of the output files (plots)}

\item{sf}{an \code{sf} object which will be added to the plotted maps.}

\item{settings}{xml files for plotting settings (see internal code)}

\item{fun_aggr}{character aggregation function name, used as name prefixes in namaing \code{x}'s layers  (see \code{\link{dryspellcliva}} and \code{\link{dryspellapprast}})}

\item{summary_suffixes}{suffixes used for summay/regression functions (see \code{\link{regress}})}

\item{...}{further arguments passed to \code{\link{ggsave}}}
}
\description{
Heat and Cold waves analysis in Spatial Gridded Coverage (visualization)
}
\examples{

library(sf)

years <- 1982:2023
dataset_path <- "/home/ecor/local/rpackages/jrc/terracliva/inst/ext_data/precipitation"
dataset_path <- system.file("ext_data/precipitation",package="terracliva")
dataset_daily <- "\%s/daily/chirps_daily_goma_\%04d.grd" \%>\% sprintf(dataset_path,years) \%>\% rast()
dataset_sf <- system.file("ext_data/OSM_Goma_quartiers_210527.shp",package="terracliva") \%>\% st_read()
filename_dry_spell <- "/home/ecor/local/rpackages/jrc/terraclivaviz/inst/ext_data/goma_dry_spell.tif"

fun_aggr=aggr_fun_suffixes()
if (!file.exists(filename_dry_spell)) {
  out <- dryspellapprast(dataset_daily,valmin=1,filename=filename_dry_spell,summary_regress=TRUE,fun_aggr=fun_aggr,overwrite=TRUE)
} else {
  out <- rast(filename_dry_spell)
}



filenames <- "/home/ecor/local/rpackages/jrc/terraclivaviz/inst/examples/plot/dryspell/dryspell_\%s.jpg"
out2 <- dryspellapprastviz(out,sf=dataset_sf,filenames=filenames)



}
