% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwmidapprastviz.R
\name{hwmidapprastviz}
\alias{hwmidapprastviz}
\title{Heat and Cold waves analysis in Spatial Gridded Coverage (visualization)}
\usage{
hwmidapprastviz(
  x,
  filenames,
  sf,
  settings = system.file("settings/lm_plot_settings_enexus.xml", package =
    "terraclivaviz"),
  mask = FALSE,
  write_tif = FALSE,
  nrow_all = NULL,
  ncol_all = 6,
  width = NA,
  height = NA,
  width_panel = width,
  height_panel = height,
  limitsize = FALSE,
  dpi = 300,
  units = "mm",
  create.dir = TRUE,
  add_comprehensive_view = TRUE,
  ...
)
}
\arguments{
\item{x}{a \code{SpatRast-Class} object returned by \code{\link{hwmidapprast}}.}

\item{filenames}{vector or string for names of the output files (plots)}

\item{sf}{an \code{sf} object which will be added to the plotted maps.}

\item{settings}{xml files for plotting settings (see internal code)}

\item{mask}{logical If it is \code{TRUE} only the area within the \code{sf} shape is visualized. Default is \code{FALSE}}

\item{write_tif}{logical. Default is \code{FALSE}. If \code{TRUE}, results are also written and saved as GeoTiff raster files.}

\item{nrow_all, ncol_all}{number of rows and columns for frames for the comprehensive panel}

\item{width, height, limitsize, dpi, units, create.dir, ...}{further arguments passed to \code{\link[ggplot2]{ggsave}} (see defailt values in function usage)}

\item{width_panel, height_panel}{width and height of a single panal within a comprehensive (frame) plot.}

\item{add_comprehensive_view}{logical . If \code{TRUE} a comprehensive plot with panels of raster maps is presented.}
}
\description{
Heat and Cold waves analysis in Spatial Gridded Coverage (visualization)
}
\note{
\code{x} must have the proper time aggregation for the analysis before the execution of this function.
}
\examples{

library(sf)
years <- 1983:2016
tmax_dataset_path <- system.file("ext_data/tmax",package="terracliva")
tmax_dataset_daily <- "\%s/daily/chirts_daily_goma_tmax_\%04d.grd" \%>\% 
sprintf(tmax_dataset_path,years) \%>\% rast()
dataset_sf <- system.file("ext_data/OSM_Goma_quartiers_210527.shp"
,package="terracliva") \%>\% st_read()


o_hw_regress <- hwmidapprast(tmax_dataset_daily,summary_regress=TRUE)

filenames <- system.file(package="terraclivaviz") \%>\% 
  file.path("examples/plot/hwmid/hwmid_\%s.jpg")
###filenames <- "/home/ecor/local/rpackages/jrc/terraclivaviz_examples/examples/plot/hwmid/hwmid_\%s.jpg"
out_hw_regress_viz <- hwmidapprastviz(x=o_hw_regress,filenames,sf=dataset_sf)

filenames <- system.file(package="terraclivaviz") \%>\% 
   file.path("examples/plot/hwmid/hwmid_svg_printing_\%s.svg")

out_hw_regress_svg_viz <- hwmidapprastviz(x=o_hw_regress,filenames,sf=dataset_sf)


}
