% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiapprastviz.R
\name{spiapprastviz}
\alias{spiapprastviz}
\title{SPI / SPEI index  (see lmomPi implementation)  Climate Variability Analysis in Spatial Gridded Coverage (visualization)}
\usage{
spiapprastviz(
  x,
  filenames,
  sf,
  settings = system.file("settings/lm_plot_settings_enexus.xml", package =
    "terraclivaviz"),
  signif = attr(x, "signif"),
  mask = FALSE,
  write_tif = FALSE,
  ...
)
}
\arguments{
\item{x}{a \code{SpatRast-Class} object returned by \code{\link{siapprast}}.}

\item{filenames}{vector or string for names of the output files (plots)}

\item{sf}{an \code{sf} object which will be added to the plotted maps.}

\item{settings}{xml files for plotting settings (see internal code)}

\item{mask}{logical If it is \code{TRUE} only the area within the \code{sf} shape is visualized. Default is \code{FALSE}}

\item{write_tif}{logical. Default is \code{FALSE}. If \code{TRUE}, results are also written and saved as GeoTiff raster files.}

\item{...}{further arguments passed to \code{\link{ggsave}}}
}
\description{
SPI / SPEI index  (see lmomPi implementation)  Climate Variability Analysis in Spatial Gridded Coverage (visualization)
}
\note{
\code{x} must have the proper time aggregation for the analysis before the execution of this function.
}
\examples{

library(magrittr)
library(terra)
library(lmomPi)
library(sf)

years <- 1982:2023

dataset_path <- system.file("ext_data/precipitation",package="terracliva")
dataset_monthly <- "\%s/monthly/chirps_monthly_goma_\%04d.grd" \%>\% 
sprintf(dataset_path,years) \%>\% rast()
terra::time(dataset_monthly) <-  names(dataset_monthly) \%>\% 
paste0("_01") \%>\% as.Date(format="X\%Y_\%m_\%d")
dataset_sf <- system.file("ext_data/OSM_Goma_quartiers_210527.shp"
,package="terracliva") \%>\% st_read()

o_spi1t <- spiapprast(x=dataset_monthly,distrib="pe3",summary_regress=TRUE)

###filenames <- system.file(package="terraclivaviz") \%>\% file.path("examples/plot/spi/spi_\%s.jpg")
filenames <- "/home/ecor/local/rpackages/jrc/terraclivaviz/inst/examples/plot/spi/spi_\%s.jpg"
out_spi1t_viz <- spiapprastviz(x=o_spi1t,filenames,sf=dataset_sf)

###filenames <- system.file(package="terraclivaviz") \%>\% file.path("examples/plot/spi/spi_svg_printing_\%s.svg")
filenames <- "/home/ecor/local/rpackages/jrc/terraclivaviz/inst/examples/plot/spi/spi_svg_printing_\%s.svg"
out_spi1t_viz <- spiapprastviz(x=o_spi1t,filenames,sf=dataset_sf,write_tif=TRUE)





}
