% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryspellapprastviz.R
\name{dryspellapprastviz}
\alias{dryspellapprastviz}
\title{Dry Spell Analysis in Spatial Gridded Coverage (visualization)}
\usage{
dryspellapprastviz(
  x,
  filenames,
  sf,
  settings = system.file("settings/lm_plot_settings_enexus.xml", package =
    "terraclivaviz"),
  fun_aggr = terracliva::aggr_fun_suffixes(),
  summary_suffixes = c("pvalue", "coeff", "stdrerror", "rsquared", "senslope",
    "pvalue_mk"),
  mask = FALSE,
  write_tif = FALSE,
  create.dir = TRUE,
  ...
)
}
\arguments{
\item{x}{a \code{SpatRast-Class} object returned by \code{\link{dryspellapprast}}.}

\item{filenames}{vector or string for names of the output files (plots)}

\item{sf}{an \code{sf} object which will be added to the plotted maps.}

\item{settings}{xml files for plotting settings (see internal code)}

\item{fun_aggr}{character aggregation function name, used as name prefixes in namaing \code{x}'s layers  (see \code{\link{dryspellcliva}} and \code{\link{dryspellapprast}})}

\item{summary_suffixes}{suffixes used for summary/regression functions (see \code{\link{regress}})}

\item{mask}{logical If it is \code{TRUE} only the area within the \code{sf} shape is visualized. Default is \code{FALSE}}

\item{write_tif}{logical. Default is \code{FALSE}. If \code{TRUE}, results are also written and saved as GeoTiff raster files.}

\item{create.dir, ...}{further arguments passed to \code{\link[ggplot2]{ggsave}}}
}
\description{
Dry Spell Analysis in Spatial Gridded Coverage (visualization)
}
\examples{

library(sf)

years <- 1982:2023
dataset_path <- system.file("ext_data/precipitation",package="terracliva")
dataset_daily <- "\%s/daily/chirps_daily_goma_\%04d.grd" \%>\% sprintf(dataset_path,years) \%>\% rast()
dataset_sf <- system.file(
"ext_data/OSM_Goma_quartiers_210527.shp",package="terracliva") \%>\% 
st_read()
filename_dry_spell <- system.file("ext_data/goma_dry_spell.tif",package="terraclivaviz")

fun_aggr=aggr_fun_suffixes()
if (!file.exists(filename_dry_spell)) {
  out <- dryspellapprast(dataset_daily,
  valmin=1,filename=filename_dry_spell,
  summary_regress=TRUE,fun_aggr=fun_aggr,overwrite=TRUE)
} else {
  out <- rast(filename_dry_spell)
}



filedir <- system.file(package="terraclivaviz") \%>\% file.path("examples/plot/dryspell/")
filenames <- paste0(filedir,"dryspell_\%s.jpg")
out2 <- dryspellapprastviz(out,sf=dataset_sf,filenames=filenames,write_tif=TRUE)

uu <- list.files(filedir,full.names=TRUE)

}
