% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwmidapprastviz.R
\name{hwmidapprastviz}
\alias{hwmidapprastviz}
\title{Heat and Cold waves analysis in Spatial Gridded Coverage (visualization)}
\usage{
hwmidapprastviz(
  x,
  filenames,
  sf,
  settings = system.file("settings/lm_plot_settings_v4.xml", package = "terraclivaviz"),
  ...
)
}
\arguments{
\item{x}{a \code{SpatRast-Class} object returned by \code{\link{hwmidapprast}}.}

\item{filenames}{vector or string for names of the output files (plots)}

\item{sf}{an \code{sf} object which will be added to the plotted maps.}

\item{settings}{xml files for plotting settings (see internal code)}

\item{...}{further arguments passed to \code{\link{ggsave}}}
}
\description{
Heat and Cold waves analysis in Spatial Gridded Coverage (visualization)
}
\note{
\code{x} must have the proper time aggregation for the analysis before the execution of this function.
}
\examples{

library(sf)
years <- 1983:2016
tmax_dataset_path <- system.file("ext_data/tmax",package="terracliva")
tmax_dataset_daily <- "\%s/daily/chirts_daily_goma_tmax_\%04d.grd" \%>\% sprintf(tmax_dataset_path,years) \%>\% rast()
dataset_sf <- system.file("ext_data/OSM_Goma_quartiers_210527.shp",package="terracliva") \%>\% st_read()


o_hw_regress <- hwmidapprast(tmax_dataset_daily,summary_regress=TRUE)

filenames <- "/home/ecor/local/rpackages/jrc/terraclivaviz/inst/examples/plot/hwmid/hwmid_\%s.jpg"
out_hw_regress_viz <- hwmidapprastviz(x=o_hw_regress,filenames,sf=dataset_sf)

}
