% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmapprastviz.R
\name{lmapprastviz}
\alias{lmapprastviz}
\title{Precipitation Deficit with L-Moments  Climate Variability Analysis in Spatial Gridded Coverage (visualization)}
\usage{
lmapprastviz(
  x,
  filenames,
  sf,
  distrib = eval(formals(lmomPi::pel)$distrib),
  settings = system.file("settings/lm_plot_settings_enexus.xml", package =
    "terraclivaviz"),
  mask = FALSE,
  write_tif = FALSE,
  use_levelplot = FALSE,
  use_ggplot2 = !use_levelplot,
  device = "png",
  ...
)
}
\arguments{
\item{x}{a \code{SpatRast-Class} object returned by \code{\link{lmapprast}}.}

\item{filenames}{vector or string for names of the output files (plots)}

\item{sf}{an \code{sf} object which will be added to the plotted maps.}

\item{distrib}{see \code{\link{pel}}}

\item{settings}{xml files for plotting settings (see internal code)}

\item{mask}{logical If it is \code{TRUE} only the area within the \code{sf} shape is visualized. Default is \code{FALSE}}

\item{write_tif}{logical. Default is \code{FALSE}. If \code{TRUE}, results are also written and saved as GeoTiff raster files.}

\item{use_levelplot}{logical (experimantal). If \code{TRUE}  plots ara made with \code{\link{levelplot}}}

\item{use_ggplot2}{logical. If \code{TRUE} (default) plots ara mede with \code{ggplot2}}

\item{device}{used if \code{use_levelplot==TRUE} , argument paased to passed to \code{\link{trellis.device}}}

\item{...}{further arguments passed to \code{\link{ggsave}} or alternatively \code{\link{trellis.device}}}
}
\description{
Precipitation Deficit with L-Moments  Climate Variability Analysis in Spatial Gridded Coverage (visualization)
}
\note{
\code{x} must have the proper time aggregation for the analysis before the execution of this function.
}
\examples{


library(sf)

years <- 1982:2023



dataset_path <- system.file("ext_data/precipitation",package="terracliva")
dataset_yearly <- "\%s/yearly/chirps_yearly_goma_\%04d.grd" \%>\% sprintf(dataset_path,years) \%>\% rast()
dataset_sf <- system.file("ext_data/OSM_Goma_quartiers_210527.shp",package="terracliva") \%>\% 
  st_read()

out_yearly <- lmapprast(dataset_yearly)

filenames <- system.file(package="terraclivaviz") \%>\% 
  file.path("examples/plot/lm/yearly/lm_\%s.jpg")
out_yearly_viz <- lmapprastviz(x=out_yearly,filenames,sf=dataset_sf)




library(lubridate)
dataset_monthly <- "\%s/monthly/chirps_monthly_goma_\%04d.grd" \%>\% 
  sprintf(dataset_path,years) \%>\% rast()
time(dataset_monthly) <-  names(dataset_monthly) \%>\% paste0("_01") \%>\% as.Date(format="X\%Y_\%m_\%d")


out_monthly <- lmapprast(dataset_monthly,index="monthly",distrib="pe3")

filenames <- system.file(package="terraclivaviz") \%>\% 
  file.path("examples/plot/lm/monthly/lm_\%s.jpg")
out_monthly_viz <- lmapprastviz(x=out_monthly,filenames,sf=dataset_sf)


out_monthly_viz <- lmapprastviz(x=out_monthly,filenames,sf=dataset_sf,use_levelplot=TRUE)


}
