% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiapprastviz.R
\name{spiapprastviz}
\alias{spiapprastviz}
\title{SPI / SPEI index  (see lmomPi implementation)  Climate Variability Analysis in Spatial Gridded Coverage (visualization)}
\usage{
spiapprastviz(
  x,
  filenames,
  sf,
  settings = system.file("settings/lm_plot_settings_enexus.xml", package =
    "terraclivaviz"),
  signif = attr(x, "signif"),
  mask = FALSE,
  write_tif = FALSE,
  add_spatial_statistics = !is.null(attr(x, "spi_cat")),
  id.name = "NAME",
  month = 1:12,
  ...
)
}
\arguments{
\item{x}{a \code{SpatRast-Class} object returned by \code{\link{spiapprast}}.}

\item{filenames}{vector or string for names of the output files (plots) .}

\item{sf}{an \code{sf} object which will be added to the plotted maps.}

\item{settings}{xml files for plotting settings (see internal code).}

\item{signif}{significance , used to detect trend occurrence, see \code{\link{spiapprast}}.}

\item{mask}{logical If it is \code{TRUE} only the area within the \code{sf} shape is visualized. Default is \code{FALSE}.}

\item{write_tif}{logical. Default is \code{FALSE}. If \code{TRUE}, results are also written and saved as GeoTiff raster files.}

\item{add_spatial_statistics}{logical. If \code{TRUE} spatial statistics are calculated on the areas/polygons of \code{sf} geospatial vector object.}

\item{id.name}{id name in \code{sf} for each areas/polygons.}

\item{month}{months to be selected for spatial stastics.}

\item{...}{further arguments passed to \code{\link{ggsave}}}
}
\description{
SPI / SPEI index  (see lmomPi implementation)  Climate Variability Analysis in Spatial Gridded Coverage (visualization)
}
\note{
\code{x} must have the proper time aggregation for the analysis before the execution of this function.
}
\examples{

library(magrittr)
library(terracliva)
library(lmomPi)
library(sf)


years <- 1982:2023
dataset_monthly <- system.file("ext_data/mekrou/CHIRPS_StackMekrou.nc",package="terraclivaviz")
dataset_monthly <- rast(dataset_monthly)+0


dataset_sf <- system.file("ext_data/mekrou/Mekrou_AOI/Mekrou_AOI_v3.shp",
package="terraclivaviz") \%>\% 
 st_read()

outspi <- spiapprast(x=dataset_monthly,distrib="pe3",summary_regress=TRUE,add_cat=TRUE,spi.scale=1)


filenames <- system.file(package="terraclivaviz") \%>\% 
 file.path("examples/plot/spi/spi_mekrow_\%s.jpg")
out <- spiapprastviz(x=outspi,filenames=filenames,
 sf=dataset_sf,signif=0.1,
 write_tif=TRUE,month=6:10)


}
