% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiapprastviz.R
\name{spiapprastviz}
\alias{spiapprastviz}
\title{SPI / SPEI index  (see lmomPi implementation)  Climate Variability Analysis in Spatial Gridded Coverage (visualization)}
\usage{
spiapprastviz(
  x,
  filenames,
  sf,
  settings = system.file("settings/lm_plot_settings_enexus.xml", package =
    "terraclivaviz"),
  signif = attr(x, "signif"),
  mask = FALSE,
  write_tif = FALSE,
  add_spatial_statistics = !is.null(attr(x, "spi_cat")),
  add_spi_cat_maps = add_spatial_statistics,
  id.name = "NAME",
  sel_regions = NA,
  month = 1:12,
  add_comprehensive_view = TRUE,
  nrow_all = NULL,
  ncol_all = length(month),
  width = NA,
  height = NA,
  width_panel = width,
  height_panel = height,
  limitsize = FALSE,
  dpi = 300,
  units = "mm",
  create.dir = TRUE,
  write_shp = TRUE,
  spi.scale = 1,
  spi.classes = read.table(system.file("settings/spi_class.csv", package = "terracliva"),
    header = TRUE, sep = ",", comment.char = "?"),
  ...
)
}
\arguments{
\item{x}{a \code{SpatRast-Class} object returned by \code{\link{spiapprast}}.}

\item{filenames}{vector or string for names of the output files (plots) .}

\item{sf}{an \code{sf} object which will be added to the plotted maps.}

\item{settings}{xml files for plotting settings (see internal code).}

\item{signif}{significance , used to detect trend occurrence, see \code{\link{spiapprast}}.}

\item{mask}{logical If it is \code{TRUE} only the area within the \code{sf} shape is visualized. Default is \code{FALSE}.}

\item{write_tif}{logical. Default is \code{FALSE}. If \code{TRUE}, results are also written and saved as GeoTiff raster files.}

\item{add_spatial_statistics}{logical. If \code{TRUE} spatial statistics are calculated on the areas/polygons of \code{sf} geospatial vector object.}

\item{add_spi_cat_maps}{logical. If \code{TRUE} spi categorical maps are displayed.}

\item{id.name}{id name in \code{sf} for each areas/polygons.}

\item{sel_regions}{selected regions/areas/polygons to display (optional), otherwise all regionsare displayed.}

\item{month}{months to be selected for spatial stastics.}

\item{add_comprehensive_view}{logical . If \code{TRUE} a comprehensive plot with panels of raster maps is presented.}

\item{nrow_all, ncol_all}{number of rows and columns for frames for the comprehensive panel}

\item{width, height, limitsize, dpi, units, create.dir, ...}{further arguments passed to \code{\link[ggplot2]{ggsave}} (see defailt values in function usage)}

\item{width_panel, height_panel}{width and height of a single panel within a comprehensive (frame) plot.}

\item{write_shp}{logical .}

\item{spi.scale}{integer value. Default is 1 , it is a property of \code{x}. See \link[terracliva]{spicliva}.}

\item{spi.classes}{data frame with SPI/SPEI classes (see default csv file)}
}
\description{
SPI / SPEI index  (see lmomPi implementation)  Climate Variability Analysis in Spatial Gridded Coverage (visualization)
}
\note{
\code{x} must have the proper time aggregation for the analysis before the execution of this function.
}
\examples{

library(magrittr)
library(terracliva)
library(sf)
library(terra)

years <- 1982:2023
dataset_monthly <- system.file("ext_data/mekrou/CHIRPS_StackMekrou.nc",package="terraclivaviz")
dataset_monthly <- rast(dataset_monthly)+0


dataset_sf <- system.file("ext_data/mekrou/Mekrou_AOI/Mekrou_AOI_v3.shp",
package="terraclivaviz") \%>\% 
 st_read()

outspi <- spiapprast(x=dataset_monthly,distrib="pe3",summary_regress=TRUE,add_cat=FALSE,spi.scale=1)


filenames <- system.file(package="terraclivaviz") \%>\% file.path("examples/plot/spi/spi_mekrou_\%s.jpg")
 
out <- spiapprastviz(x=outspi,filenames=filenames,
 sf=dataset_sf,signif=0.1,
 write_tif=TRUE,month=6:10,dpi=300,limitsize=FALSE,units="mm",add_spatial_statistics=TRUE)

filenames2 <- system.file(package="terraclivaviz") \%>\% file.path("examples/plot/spi_regions/spi_mekrou_\%s.jpg")
 
out2 <- spiapprastviz(x=outspi,filenames=filenames2,
 sf=dataset_sf,signif=0.1,
 write_tif=TRUE,month=6:10,dpi=300,limitsize=FALSE,units="mm",add_spatial_statistics=TRUE,sel_regions=c(1,2))



}
